// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  billing: typeof billing;
  billingHelpers: typeof billingHelpers;
  comments: typeof comments;
  http: typeof http;
  mux: typeof mux;
  muxActions: typeof muxActions;
  projects: typeof projects;
  s3: typeof s3;
  security: typeof security;
  shareAccess: typeof shareAccess;
  shareLinks: typeof shareLinks;
  teams: typeof teams;
  videoActions: typeof videoActions;
  videoPresence: typeof videoPresence;
  videos: typeof videos;
  workspace: typeof workspace;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;